requirements.txt
streamlit
Pillow~=8.4.0
torch~=1.10.0
transformers~=4.12.5
timm